tests/test_embeddings.py
import unittest
import numpy as np

# This is a placeholder for test cases.
# In a real project, you would write unit tests for each function
# in the core.embeddings module.

class TestEmbeddings(unittest.TestCase):
    def test_placeholder_similarity(self):
        # A simple check to ensure the logic runs without error.
        # A real test would mock the model and check outputs.
        vec1 = np.array([0.1, 0.2, 0.3])
        vec2 = np.array([0.4, 0.5, 0.6])
        # A dummy similarity calculation
        similarity = np.dot(vec1, vec2)
        self.assertIsInstance(similarity, float)

if __name__ == '__main__':
    unittest.main() 